e_commerce/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
